+++ ChangeLog.md
@@ -5,6 +5,7 @@
 * Add support for foreign-lib components
 * Make `Sha256` type abstract and add new `sha256ToByteString` deconstructor
 * Introduce `FlagName` newtype
+* Add `FromJSONKey`/`ToJSONKey` instances for `UnitId`, `PackageName`, and `PkgId`
 
 ### Executable